+++ tests/tinytest/test_funs_string_lazy.R
@@ -686,4 +686,53 @@ expect_equal_lazy(
     pull(foo)
 )
 
+# split ---------------------------------------------------------
+
+expect_equal_lazy(
+  mutate(test, foo = str_split(x8, "-")) |>
+    pull(foo),
+  mutate(test_df, foo = str_split(x8, "-")) |>
+    pull(foo)
+)
+
+expect_warning(
+  mutate(test, foo = str_split(x8, "-", n = 2)) |>
+    pull(foo),
+  "will not be used"
+)
+
+expect_warning(
+  mutate(test, foo = str_split(x8, "-", simplify = TRUE)) |>
+    pull(foo),
+  "will not be used"
+)
+
+# split_i ---------------------------------------------------------
+
+expect_equal_lazy(
+  mutate(test, foo = str_split_i(x8, "-", i = 3)) |>
+    pull(foo),
+  mutate(test_df, foo = str_split_i(x8, "-", i = 3)) |>
+    pull(foo)
+)
+
+expect_equal_lazy(
+  mutate(test, foo = str_split_i(x8, "-", i = 100)) |>
+    pull(foo),
+  mutate(test_df, foo = str_split_i(x8, "-", i = 100)) |>
+    pull(foo)
+)
+
+expect_equal_lazy(
+  mutate(test, foo = str_split_i(x8, "-", i = -1)) |>
+    pull(foo),
+  mutate(test_df, foo = str_split_i(x8, "-", i = -1)) |>
+    pull(foo)
+)
+
+expect_error_lazy(
+  mutate(test, foo = str_split_i(x8, "-", i = 0)),
+  "must not be 0"
+)
+
 Sys.setenv('TIDYPOLARS_TEST' = FALSE)