+++ NEWS.md
@@ -10,7 +10,7 @@
   - from package `dplyr`: `consecutive_id()`, `min_rank()`, `na_if()`, 
     `n_distinct()`, `nth()`
   
-  - from package `stringr`: `str_dup()`
+  - from package `stringr`: `str_dup()`, `str_split()`, `str_split_i()`
 
 ## Bug fixes